Update v3.7.2
Added useKey icon back to doorOverlay (mistakenly deleted it)
Keys can now be named and will share what keys were used when unlocking the door
Removed all on_ macros and include Default Macro and Default Sounds in the config menu.
Default Macros and sounds will be ignored if same values set on the doors themselves.
The version AFTER this will support migration of APinterrupt and APsounds
APinterrupt will now be run OFTEN and can be used to customize ActionPad behaviors and/or integrate with other tokens and add-ons
More to come with APinterrupt . . .